+++ performance/suffix_array_construction.c
@@ -8,10 +8,10 @@
 #include <string.h>
 #include <assert.h>
 
-static char *build_equal(size_t size)
+static char *build_equal(uint32_t size)
 {
     char *s = malloc(size + 1);
-    for (size_t i = 0; i < size; ++i) {
+    for (uint32_t i = 0; i < size; ++i) {
         s[i] = 'A';
     }
     s[size] = '\0';
@@ -19,13 +19,13 @@ static char *build_equal(size_t size)
     return s;
 }
 
-static char *build_random(size_t size)
+static char *build_random(uint32_t size)
 {
     const char *alphabet = "ACGT";
     int n = strlen(alphabet);
     char *s = malloc(size + 1);
     
-    for (size_t i = 0; i < size; ++i) {
+    for (uint32_t i = 0; i < size; ++i) {
         s[i] = alphabet[rand() % n];
     }
     s[size] = '\0';
@@ -35,7 +35,7 @@ static char *build_random(size_t size)
 
 
 
-static void get_performance(size_t size)
+static void get_performance(uint32_t size)
 {
     char *s;
     struct suffix_array *sa;
@@ -46,19 +46,19 @@ static void get_performance(size_t size)
     begin = clock();
     sa = qsort_sa_construction(s);
     end = clock();
-    printf("Quick-sort Equal %lu %f\n", size, (double)(end - begin) / CLOCKS_PER_SEC);
+    printf("Quick-sort Equal %u %f\n", size, (double)(end - begin) / CLOCKS_PER_SEC);
     free_suffix_array(sa);
     
     begin = clock();
     sa = skew_sa_construction(s);
     end = clock();
-    printf("Skew_v5 Equal %lu %f\n", size, (double)(end - begin) / CLOCKS_PER_SEC);
+    printf("Skew_v5 Equal %u %f\n", size, (double)(end - begin) / CLOCKS_PER_SEC);
     free_suffix_array(sa);
     
     begin = clock();
     sa = skew_sa_construction(s);
     end = clock();
-    printf("Skew_v5 Equal %lu %f\n", size, (double)(end - begin) / CLOCKS_PER_SEC);
+    printf("Skew_v5 Equal %u %f\n", size, (double)(end - begin) / CLOCKS_PER_SEC);
     free_suffix_array(sa);
 
     free(s);
@@ -67,19 +67,19 @@ static void get_performance(size_t size)
     begin = clock();
     sa = qsort_sa_construction(s);
     end = clock();
-    printf("Quick-sort Random %lu %f\n", size, (double)(end - begin) / CLOCKS_PER_SEC);
+    printf("Quick-sort Random %u %f\n", size, (double)(end - begin) / CLOCKS_PER_SEC);
     free_suffix_array(sa);
 
     begin = clock();
     sa = skew_sa_construction(s);
     end = clock();
-    printf("Skew_v5 Random %lu %f\n", size, (double)(end - begin) / CLOCKS_PER_SEC);
+    printf("Skew_v5 Random %u %f\n", size, (double)(end - begin) / CLOCKS_PER_SEC);
     free_suffix_array(sa);
 
     begin = clock();
     sa = skew_sa_construction(s);
     end = clock();
-    printf("Skew_v5 Random %lu %f\n", size, (double)(end - begin) / CLOCKS_PER_SEC);
+    printf("Skew_v5 Random %u %f\n", size, (double)(end - begin) / CLOCKS_PER_SEC);
     free_suffix_array(sa);
 
     free(s);
@@ -103,7 +103,7 @@ int main(int argc, const char **argv)
 {
     srand(time(NULL));
     
-    for (size_t n = 0; n < 10000; n += 500) {
+    for (uint32_t n = 0; n < 10000; n += 500) {
         for (int rep = 0; rep < 5; ++rep) {
             get_performance(n);
         }
